Move test-adapter.ts to /query-testing/test-adapter.ts
