+++ extensions/ql-vscode/src/query-testing/test-adapter.ts
@@ -18,12 +18,12 @@ import {
   QLTestNode,
   QLTestDirectory,
   QLTestDiscovery,
-} from "./qltest-discovery";
+} from "../qltest-discovery";
 import { Event, EventEmitter, CancellationTokenSource } from "vscode";
-import { DisposableObject } from "./pure/disposable-object";
-import { CodeQLCliServer, TestCompleted } from "./codeql-cli/cli";
-import { testLogger } from "./common";
-import { TestRunner } from "./test-runner";
+import { DisposableObject } from "../pure/disposable-object";
+import { CodeQLCliServer, TestCompleted } from "../codeql-cli/cli";
+import { testLogger } from "../common";
+import { TestRunner } from "../test-runner";
 
 /**
  * Get the full path of the `.expected` file for the specified QL test.
