+++ extensions/ql-vscode/src/test-manager-base.ts
@@ -1,7 +1,7 @@
 import { copy, createFile, lstat, pathExists } from "fs-extra";
 import { TestUICommands } from "./common/commands";
 import { DisposableObject } from "./pure/disposable-object";
-import { getActualFile, getExpectedFile } from "./test-adapter";
+import { getActualFile, getExpectedFile } from "./query-testing/test-adapter";
 import { TestItem, TextDocumentShowOptions, Uri, window } from "vscode";
 import { basename } from "path";
 import { App } from "./common/app";
