+++ extensions/ql-vscode/src/test-ui.ts
@@ -9,7 +9,7 @@ import {
 } from "vscode-test-adapter-api";
 import { TestTreeNode } from "./test-tree-node";
 import { DisposableObject } from "./pure/disposable-object";
-import { QLTestAdapter } from "./test-adapter";
+import { QLTestAdapter } from "./query-testing/test-adapter";
 import { App } from "./common/app";
 import { TestManagerBase } from "./test-manager-base";
