+++ extensions/ql-vscode/test/vscode-tests/no-workspace/query-testing/test-adapter.test.ts
@@ -10,18 +10,18 @@ import {
   tests,
 } from "vscode";
 
-import { QLTestAdapter } from "../../../src/test-adapter";
-import { CodeQLCliServer } from "../../../src/codeql-cli/cli";
-import { DatabaseManager } from "../../../src/databases/local-databases";
-import { mockedObject } from "../utils/mocking.helpers";
-import { TestRunner } from "../../../src/test-runner";
+import { QLTestAdapter } from "../../../../src/query-testing/test-adapter";
+import { CodeQLCliServer } from "../../../../src/codeql-cli/cli";
+import { DatabaseManager } from "../../../../src/databases/local-databases";
+import { mockedObject } from "../../utils/mocking.helpers";
+import { TestRunner } from "../../../../src/test-runner";
 import {
   createMockCliServerForTestRun,
   mockEmptyDatabaseManager,
   mockTestsInfo,
-} from "./test-runner-helpers";
-import { TestManager } from "../../../src/test-manager";
-import { createMockApp } from "../../__mocks__/appMock";
+} from "../test-runner-helpers";
+import { TestManager } from "../../../../src/test-manager";
+import { createMockApp } from "../../../__mocks__/appMock";
 
 type IdTestItemPair = [id: string, testItem: TestItem];
